Add authentication middleware to routes
