+++ lib/util.js
@@ -2,6 +2,8 @@
  * Util.js
  *
  */
+ var OAuth = require('oauth').OAuth2;
+ var config = CONFIG;
 
  exports.sortObject = function (obj) {
   var tempArray = [];
@@ -21,3 +23,33 @@
 
   return tempObj;
  };
+
+
+ /**
+  * Authentication filter.
+  * Used as a middleware on all routes.
+  */
+
+  exports.isAuthenticated = function (req, res, next) {
+    if (req.signedCookies.accessToken) {
+      return next();
+    } else {
+      var auth = new OAuth(
+        config.githubClientId,
+        config.githubClientSecret,
+        'https://github.com/',
+        'login/oauth/authorize',
+        'login/oauth/access_token',
+        null
+      );
+
+      /* Get the authorize URL from Github */
+      var authUrl = auth.getAuthorizeUrl({
+        redirect_uri: process.env.GITHUB_OAUTH_REDIRECT_URI + '/login',
+        scope: ['repo'],
+        state: config.githubState
+      });
+
+      res.render('index', {authUrl: authUrl});
+    }
+  };
